Rename test-tools/ to test-tool/ as other dirs are singular
