+++ src/command/test/DebugRepo.spec.ts
@@ -1,6 +1,6 @@
-import AppTestHarness from 'test-tools/helpers/AppTestHarness';
+import AppTestHarness from 'test-tool/helpers/AppTestHarness';
 import DebugRepo from 'command/DebugRepo';
-import getContextStub from 'test-tools/helpers/getContextStub';
+import getContextStub from 'test-tool/helpers/getContextStub';
 import JobProviderFactory from 'job/JobProviderFactory';
 
 let debugRepo: DebugRepo;
