+++ src/common/test/Workspace.spec.ts
@@ -1,4 +1,4 @@
-import AppTestHarness from 'test-tools/helpers/AppTestHarness';
+import AppTestHarness from 'test-tool/helpers/AppTestHarness';
 import FakeEditorGateway from 'gateway/FakeEditorGateway';
 import Workspace from 'common/Workspace';
