+++ src/containerization/test/Docker.spec.ts
@@ -1,4 +1,4 @@
-import AppTestHarness from 'test-tools/helpers/AppTestHarness';
+import AppTestHarness from 'test-tool/helpers/AppTestHarness';
 import ChildProcessGateway from 'gateway/ChildProcessGateway';
 import Docker from 'containerization/Docker';
 import Types from 'common/Types';
