+++ src/test-tool/expected/dynamic-config.yml
